+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2026-01-22
+
+- Provide a new method `get_level_shape` to help get the shape after decoding for
+  the image at a given resolution level.
+
 ## [0.5.1] - 2025-12-29
 
 - Fix writing 3D arrays with a single channel dimension (shape `(H, W, 1)`) as monochrome